requirements.txt
django==1.5.1
south==0.8.1
lxml==3.2.3
django-registration==1.0
django-taggit==0.10a1
dj-static==0.0.5
django-extensions==1.1.1
requests==1.2.3
